+++ src/starknet/hash.rs
@@ -3,3 +3,4 @@ use serde::{Deserialize, Serialize};
     Debug, Copy, Clone, PartialEq, Eq, Default, Hash, Deserialize, Serialize, PartialOrd, Ord,
 )]
 pub struct StarkHash(pub [u8; 32]);
+pub type StarkFelt = StarkHash;
