+++ src/starknet/mod.rs
@@ -5,5 +5,4 @@ pub use block::{
     BlockBody, BlockHash, BlockHeader, BlockNumber, BlockTimestamp, ContractAddress,
     EventsCommitment, GasPrice, GlobalRoot, TransactionsCommitment,
 };
-
-pub use hash::StarkHash;
+pub use hash::{StarkFelt, StarkHash};
