+++ src/sync/mod.rs
@@ -6,9 +6,10 @@ use futures_util::pin_mut;
 use log::{debug, error, info};
 use tokio_stream::StreamExt;
 
-use crate::{
-    starknet_client::ClientError,
-    storage::components::{BlockStorageError, BlockStorageReader, BlockStorageWriter},
+use crate::starknet_client::ClientError;
+use crate::storage::components::{
+    BlockStorageError, BlockStorageReader, BlockStorageWriter, HeaderStorageReader,
+    HeaderStorageWriter,
 };
 
 pub use self::sources::CentralSource;
